requirements.txt
requests
uiautomation
loguru
python-dotenv
